+++ SuperCyanTweaks/Enemies/VoidJailer.cs
@@ -8,6 +8,7 @@ namespace SuperCyanTweaks
     public class VoidJailer
     {
         public static GameObject jailerMasterPrefab = Addressables.LoadAssetAsync<GameObject>("RoR2/DLC1/VoidJailer/VoidJailerMaster.prefab").WaitForCompletion();
+        public static GameObject jailerAllyMasterPrefab = Addressables.LoadAssetAsync<GameObject>("RoR2/DLC1/VoidJailer/VoidJailerAllyMaster.prefab").WaitForCompletion();
 
         public VoidJailer()
         {
@@ -15,6 +16,15 @@ namespace SuperCyanTweaks
             if (Configs.voidJailerIdleFix.Value == true)
             {
                 AISkillDriver[] skillDrivers = jailerMasterPrefab.GetComponents<AISkillDriver>();
+                foreach (AISkillDriver skillDriver in skillDrivers)
+                {
+                    if (skillDriver.customName == "PathFromAfar")
+                    {
+                        skillDriver.selectionRequiresTargetLoS = false;
+                    }
+                }
+
+                skillDrivers = jailerAllyMasterPrefab.GetComponents<AISkillDriver>();
                 foreach (AISkillDriver skillDriver in skillDrivers)
                 {
                     if (skillDriver.customName == "PathFromAfar")